Overhaul the API and performance.
Insteafd of requiring users to clone their evacuation future by using
`Shared`, `Evacuate` itself now supports being cloned by introducing
some background machinery that manages the count/timeout logic as well
as notifying tasks polling an evacuate future.

As this requires a separate task to drive this behavior, `Evacuate::new`
now returns a third value in the tuple, a future that must be spawned on
an available executor.  There's also a new function,
`Evacuate::default_executor` which will spawn the runner for the caller
on the default Tokio executor.  This means it can only be used when
already instead a running task.